Improve module augmentation for `media-embed` plugin
