+++ packages/ckeditor5-media-embed/src/mediaembedtoolbar.ts
@@ -49,9 +49,3 @@ export default class MediaEmbedToolbar extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ MediaEmbedToolbar.pluginName ]: MediaEmbedToolbar;
-	}
-}
